test: structural improvements — split test_data_object.py, add conftest, fix anti-patterns
- Split monolithic test_data_object.py (388 lines) into focused files:
  - test_init.py: constructor, strict/non-strict, __call__
  - test_access.py: __getattr__, __setitem__, __setattr__, get, dir, str/repr
  - test_nesting.py: nested DOs, nullable nested, compile-time restriction validation
- Add tests/conftest.py with shared abc_cleanup fixture
- Update test_abc_restrictions.py to use shared fixture (replaces manual teardown)
- Remove MyTestException from data.py
- Replace all try/except anti-patterns with pytest.raises
- Add assertion messages to bare assert statements

181 passed, 76 xfailed
